- Removed dual Mode - Removed ununsed speed-related stuff - Moved (H)DfsBroadcast to its own file
